requirements.txt
conan
numpy
opencv-python
pandas
pyaudio
pynput
pyserial
PySide6
